Stop exposing performance-oriented FromPgField methods
So we can change them
